make CppBase more flexible
instead of _requiring_ that the module be defined in _one_ .cpp and .h
file, only require the binding_*.cpp, but allow additional source files
ot be added to the compilation unit. The header files will be included
automatically.

If necessary, adding the possibility for further include folders later
should be trivial.